+++ Changelog.md
@@ -60,6 +60,7 @@
 * Add `parse_hex_string` for 3 digit colors
 * Add parsing 'DocxParagraph#inserted' context
 * Add parsing `bgPr` and `stretch`
+* Add parsing `DisplayLabelsProperties` in `Series`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit